+++ packages/video_player_avplay/tizen/src/media_player.cc
@@ -86,8 +86,8 @@ int64_t MediaPlayer::Create(const std::string &uri,
     return -1;
   }
 
-  std::string cookie = flutter_common::GetValue(create_message.http_headers(),
-                                                "Cookie", std::string());
+  std::string cookie = flutter_common::GetValue(
+      create_message.streaming_property(), "COOKIE", std::string());
   if (!cookie.empty()) {
     int ret =
         player_set_streaming_cookie(player_, cookie.c_str(), cookie.size());
@@ -97,7 +97,7 @@ int64_t MediaPlayer::Create(const std::string &uri,
     }
   }
   std::string user_agent = flutter_common::GetValue(
-      create_message.http_headers(), "User-Agent", std::string());
+      create_message.streaming_property(), "USER_AGENT", std::string());
   if (!user_agent.empty()) {
     int ret = player_set_streaming_user_agent(player_, user_agent.c_str(),
                                               user_agent.size());