+++ packages/video_player_avplay/tizen/src/plus_player.h
@@ -41,6 +41,8 @@ class PlusPlayer : public VideoPlayer {
   bool IsReady() override;
   flutter::EncodableList GetTrackInfo(std::string track_type) override;
   bool SetTrackSelection(int32_t track_id, std::string track_type) override;
+  std::string GetStreamingProperty(
+      const std::string &streaming_property_type) override;
 
  private:
   bool IsLive();